# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Grandma's Recipes API
  description: API for managing recipes, users, and meal planning
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: Create a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
            example:
              username: "johndoe"
              firstName: "John"
              lastName: "Doe"
              country: "Israel"
              password: "Pass123!"
              email: "john@example.com"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      tags:
        - Auth
      summary: Login user
      description: Authenticate user and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCredentials'
            example:
              username: "johndoe"
              password: "Pass123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /logout:
    post:
      tags:
        - Auth
      summary: Logout user
      description: End user session
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /recipes/search:
    get:
      tags:
        - Recipes
      summary: Search recipes
      description: Search for recipes with filters
      parameters:
        - in: query
          name: query
          schema:
            type: string
          description: Search query
          example: "pasta"
        - in: query
          name: number
          schema:
            type: integer
            enum: [5, 10, 15]
            default: 5
          description: Number of results to return
        - in: query
          name: cuisine
          schema:
            type: string
          description: Filter by cuisine type
          example: "Italian"
        - in: query
          name: diet
          schema:
            type: string
          description: Filter by diet type
          example: "Vegetarian"
        - in: query
          name: intolerance
          schema:
            type: string
          description: Filter by intolerance
          example: "Gluten"
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
  /recipes/{recipeId}:
    get:
      tags:
        - Recipes
      summary: Get recipe details
      description: Get full recipe details by ID
      parameters:
        - in: path
          name: recipeId
          required: true
          schema:
            type: string
          description: Recipe ID
      security:
        - cookieAuth: []
        - {}
      responses:
        '200':
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeDetails'
        '404':
          description: Recipe not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    UserRegistration:
      type: object
      required:
        - username
        - firstName
        - lastName
        - country
        - password
        - email
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 8
          pattern: '^[a-zA-Z]+$'
          description: Username (3-8 letters only)
          example: "johndoe"
        firstName:
          type: string
          description: User's first name
          example: "John"
        lastName:
          type: string
          description: User's last name
          example: "Doe"
        country:
          type: string
          description: User's country
          example: "Israel"
        password:
          type: string
          minLength: 5
          maxLength: 10
          pattern: '^(?=.*[0-9])(?=.*[!@#$%^&*])[a-zA-Z0-9!@#$%^&*]+$'
          description: Password (5-10 chars, at least 1 number and 1 special char)
          example: "Pass123!"
        email:
          type: string
          format: email
          description: User's email
          example: "john@example.com"
    LoginCredentials:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          description: Username
          example: "johndoe"
        password:
          type: string
          description: Password
          example: "Pass123!"
    LoginResponse:
      type: object
      properties:
        message:
          type: string
          example: "Login successful"
        username:
          type: string
          example: "johndoe"
    RecipePreview:
      type: object
      properties:
        id:
          type: string
          description: Recipe ID
          example: "716429"
        title:
          type: string
          description: Recipe title
          example: "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs"
        image:
          type: string
          description: Recipe image URL
          example: "https://spoonacular.com/recipeImages/716429-556x370.jpg"
        readyInMinutes:
          type: integer
          description: Preparation time in minutes
          example: 45
        popularity:
          type: integer
          description: Recipe popularity (likes)
          example: 209
        vegetarian:
          type: boolean
          description: Is vegetarian
          example: false
        vegan:
          type: boolean
          description: Is vegan
          example: false
        glutenFree:
          type: boolean
          description: Is gluten free
          example: false
        viewed:
          type: boolean
          description: Has user viewed this recipe
          example: false
        favorite:
          type: boolean
          description: Is recipe in user's favorites
          example: false
    RecipeDetails:
      type: object
      properties:
        id:
          type: string
          description: Recipe ID
          example: "716429"
        title:
          type: string
          description: Recipe title
          example: "Pasta with Garlic, Scallions, Cauliflower & Breadcrumbs"
        image:
          type: string
          description: Recipe image URL
          example: "https://spoonacular.com/recipeImages/716429-556x370.jpg"
        readyInMinutes:
          type: integer
          description: Preparation time in minutes
          example: 45
        popularity:
          type: integer
          description: Recipe popularity (likes)
          example: 209
        vegetarian:
          type: boolean
          description: Is vegetarian
          example: false
        vegan:
          type: boolean
          description: Is vegan
          example: false
        glutenFree:
          type: boolean
          description: Is gluten free
          example: false
        servings:
          type: integer
          description: Number of servings
          example: 2
        extendedIngredients:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "flour"
              amount:
                type: number
                example: 1.5
              unit:
                type: string
                example: "cups"
        instructions:
          type: string
          description: Cooking instructions
          example: "1. Boil water..."
        viewed:
          type: boolean
          description: Has user viewed this recipe
          example: true
        favorite:
          type: boolean
          description: Is recipe in user's favorites
          example: false
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation successful"
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Error occurred"
        error:
          type: string
          example: "Invalid input"
...